requirement.txt
asgiref==3.4.1
cascade==0.1
Django==3.2.5
django-ckeditor==6.1.0
django-extensions==3.1.3
django-gravatar2==1.4.4
django-js-asset==1.2.2
djangorestframework==3.12.4
export==0.1.2
f==0.0.1
imaplib2==3.6
login==0.0.6
progressbar==2.5
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
uuid==1.30
